STLREF-master/STLREF-master/requirements.txt
streamlit
requests
pillow
numpy
gdown
